html/browsers/browsing-the-web/navigating-across-documents/004.html

<!doctype html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigating-across-documents">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src=""></iframe>
<script>
async_test(function(test) {
  var iframe = document.querySelector("iframe");
  var messages = [];

  // 004-2.html has an `onunload` handler, which sets `location` to 004-3.html.
  // This must NOT be allowed, per https://github.com/whatwg/html/commit/2a11eef8f92b0335328890dc2cb1e281d18553eb.
  //
  // 004-2.html has an `onload` handler, which sends the message "004-2" to 004.html,
  // then calls `history.go(-1)` after a 0ms timeout. `history.go(-1)` navigates
  // browsingContext to the last _session history entry_ (004-1.html).
  //
  window.onmessage = test.step_func(function(e) {
    messages.push(e.data);
    if (messages.length === 3) {
      assert_array_equals(messages, ["004-1", "004-2", "004-1"]);
      test.done();
      iframe.parentNode.removeChild(iframe);
    }
  });

  // 004-1.html has an `onload` handler, which sends the message "004-1" to
  // 004.html, then sets `location` to 004-2.html ~100ms later
  iframe.src = "004-1.html";

}, "Navigation from unload canceled when traversing cross-origin history");
</script>
